Structifying the ECS part 1.
I need to make quite a few changes to how the code works to make it compatible with structs instead of classes as components. Im making this change because due to how memory is allocated, classes are inferior to structs for large amounts of data storage. I also advise you to use efficient data types such as byte or uint or even int16 and allat so that our code is more efficient. I will be making these changes over the next coming days